new file create playlist working
