+++ backend/routes/createPlaylist.js
@@ -1,0 +1,27 @@
+import express from 'express';
+
+const router = express.Router();
+
+export default (spotifyApi) => {
+  router.post('/create-playlist', async (req, res) => {
+        try {
+        const playlistName = 'VibeFusion';
+        const options = { public: false, description: "Generated by VibeFusion" };
+        const playlistResponse = await spotifyApi.createPlaylist (playlistName, options);
+        const playlistId = playlistResponse.body.id; 
+        const { trackUris } = req.body;
+        if (trackUris.length === 0) {
+            return res.status(400).json({ error: 'Please provide an array of track URIs.' });
+        }
+    
+        const addTracksResponse = await spotifyApi.addTracksToPlaylist(playlistId, trackUris);
+    
+        res.status(200).json({message: 'yaay playlist created'}); 
+        } catch (error) {
+        res.status(500).json({ error: `Failed to create playlist ${error.message}`});
+        }
+    });
+    
+
+  return router;
+};
