+++ backend/server.js
@@ -28,6 +28,7 @@ import recommendationsRoute from './routes/recommendations.js';
 import topGenresRoute from './routes/topGenres.js';
 import profileRoute from './routes/profile.js';
 import danceabilityRoute from './routes/danceability.js';
+import createPlaylistRoute from './routes/createPlaylist.js';
 
 // Use the routes
 app.use(detailedStatsRoute(spotifyApi));
@@ -38,6 +39,7 @@ app.use(recommendationsRoute(spotifyApi));
 app.use(topGenresRoute(spotifyApi));
 app.use(profileRoute(spotifyApi));
 app.use(danceabilityRoute(spotifyApi));
+app.use(createPlaylistRoute(spotifyApi));
 
 // route for login authentication
 app.get("/login", (req, res) => {
@@ -114,26 +116,6 @@ app.get("/search", async (req, res) => {
 });
 
 
-
-// Endpoint to create a playlist named 'VibeFusion'
-app.post('/create-playlist', async (req, res) => {
-  try {
-    const playlistName = 'VibeFusion';
-    const options = { public: false, description: "Generated by VibeFusion" };
-    const playlistResponse = await spotifyApi.createPlaylist (playlistName, options);
-    const playlistId = playlistResponse.body.id; 
-    const { trackUris } = req.body;
-    if (trackUris.length === 0) {
-      return res.status(400).json({ error: 'Please provide an array of track URIs.' });
-    }
-
-    const addTracksResponse = await spotifyApi.addTracksToPlaylist(playlistId, trackUris);
-
-    res.status(200).json({message: 'yaay playlist created'}); 
-  } catch (error) {
-    res.status(500).json({ error: `Failed to create playlist ${error.message}`});
-  }
-});
